refact: move loadMissionExtras back to game.cpp
