+++ src/game.h
@@ -30,75 +30,8 @@
 #define __INCLUDED_SRC_GAME_H__
 
 #include "lib/framework/vector.h"
-
-/***************************************************************************/
-/*
- *	Global Definitions
- */
-/***************************************************************************/
-//the version number used in save games
-//#define VERSION_1				1				//demo save games
-//#define VERSION_2				2				//names saved for components/structures/features
-//#define VERSION_3				3				//changes to SAVE_GAME
-//#define VERSION_4				4				//changes to SAVE_GAME
-//#define VERSION_5				5				//different types of save game added
-//#define VERSION_6				6				//level name added to a user save game
-#define VERSION_7				7				//string ID names saved for comps an objects
-#define VERSION_8				8				//research status saved out for user saved games
-#define VERSION_9				9				//power and experience saved out for user saved games
-#define VERSION_10				10				//includes gateways and zones in .map file.
-#define VERSION_11				11				//newstyle save game with extending structure checked in 13 Nov.
-#define VERSION_12				12				//mission and order stuff checked in 20 Nov.
-//#define VERSION_13				13				//odds and ends to 24 Nov. and hashed scripts
-#define VERSION_14				14				//
-#define VERSION_15				15				//
-#define VERSION_16				16				//beta save game
-#define VERSION_17				17				//objId and new struct stats included
-#define VERSION_18				18				//droid name savegame validity stamps
-#define VERSION_19				19				//alliances, colours, radar zoom
-#define VERSION_20				20				//MAX_NAME_ SIZE replaced by MAX_SAVE_NAME_SIZE
-#define VERSION_21              21              //timeStartHold saved out for research facilities
-#define VERSION_22              22              //asRundData
-#define VERSION_23              23              //limbo droids and camstart droids saved properly, no cluster save
-#define VERSION_24              24              //reinforceTime, droid move, droid resistance
-#define VERSION_25              25              //limbo droid, research module hold fixed, cleaned by Alex
-#define VERSION_26              26              //reArm pads
-#define VERSION_27              27              //unit not the "d" word, experience and repair facility currentPtsAdded
-#define VERSION_28              28              //rearm pads currentPtsAdded saved
-#define VERSION_29              29              //mission scroll limits saved
-#define VERSION_30              30              //script external variables saved
-#define VERSION_31              31              //mission cheat time saved
-#define VERSION_32              32              //factory secondary order saved
-#define VERSION_33              33              //skirmish save
-#define VERSION_34              34              //saves AI names for multiplayer (.gam file)
-#define VERSION_35              35              //uses the (de)serialization API for saving/loading games and is big-endian instead of little-endian
-#define VERSION_36              36              //saves beacon properly
-//#define VERSION_37              37              //dpid changes; this had better be the last version
-#define VERSION_38              38              //mod list!
-#define VERSION_39		39		//lots of changes, breaking everything
-
-#define CURRENT_VERSION_NUM     VERSION_39
-
-//used in the loadGame
-#define KEEPOBJECTS				true
-#define FREEMEM					true
-
-enum GAME_TYPE
-{
-	GTYPE_SCENARIO_START,	///< Initial scenario state.
-	GTYPE_SCENARIO_EXPAND,	///< Scenario scroll area expansion.
-	GTYPE_MISSION,		///< Stand alone mission.
-	GTYPE_SAVE_START,	///< User saved game - at the start of a level.
-	GTYPE_SAVE_MIDMISSION,	///< User saved game - in the middle of a level
-};
-
-
-struct VIS_SAVEHEADER
-{
-	char        aFileType[4];
-	uint32_t    version;
-};
-
+#include "gamedef.h"
+#include "levels.h"
 
 /***************************************************************************/
 /*
@@ -111,6 +44,8 @@ bool loadGame(const char *pGameToLoad, bool keepObjects, bool freeMem, bool User
 /*This just loads up the .gam file to determine which level data to set up - split up
 so can be called in levLoadData when starting a game from a load save game*/
 bool loadGameInit(const char *fileName);
+
+bool loadMissionExtras(const char* pGameToLoad, LEVEL_TYPE levelType);
 
 // load the script state given a .gam name
 bool loadScriptState(char *pFileName);
@@ -125,8 +60,6 @@ UDWORD getCampaign(const char *fileName);
 
 /*returns the current type of save game being loaded*/
 GAME_TYPE getSaveGameType();
-
-UDWORD getSaveGameVersion();
 
 bool plotStructurePreview16(char *backDropSprite, Vector2i playeridpos[]);
