+++ src/levels.cpp
@@ -608,174 +608,6 @@ const char *getLevelName()
 	return currentLevelName;
 }
 
-
-// -----------------------------------------------------------------------------------------
-// load up saved message file
-static bool loadSaveMessage(const char* pFileName, LEVEL_TYPE levelType)
-{
-	// Only clear the messages if its a mid save game
-	if (getSaveGameType() == GTYPE_SAVE_MIDMISSION)
-	{
-		freeMessages();
-	}
-	else if (getSaveGameType() == GTYPE_SAVE_START)
-	{
-		// If we are loading in a CamStart or a CamChange then we are not interested in any saved messages
-		if (levelType == LEVEL_TYPE::LDS_CAMSTART || levelType == LEVEL_TYPE::LDS_CAMCHANGE)
-		{
-			return true;
-		}
-	}
-
-	WzConfig ini(pFileName, WzConfig::ReadOnly);
-	std::vector<WzString> list = ini.childGroups();
-	for (size_t i = 0; i < list.size(); ++i)
-	{
-		ini.beginGroup(list[i]);
-		MESSAGE_TYPE type = (MESSAGE_TYPE)ini.value("type").toInt();
-		bool bObj = ini.contains("obj/id");
-		int player = ini.value("player").toInt();
-		int id = ini.value("id").toInt();
-		int dataType = ini.value("dataType").toInt();
-
-		if (type == MSG_PROXIMITY)
-		{
-			//only load proximity if a mid-mission save game
-			if (getSaveGameType() == GTYPE_SAVE_MIDMISSION)
-			{
-				if (bObj)
-				{
-					// Proximity object so create get the obj from saved idy
-					int objId = ini.value("obj/id").toInt();
-					int objPlayer = ini.value("obj/player").toInt();
-					OBJECT_TYPE objType = (OBJECT_TYPE)ini.value("obj/type").toInt();
-					MESSAGE* psMessage = addMessage(type, true, player);
-					if (psMessage)
-					{
-						psMessage->psObj = getBaseObjFromData(objId, objPlayer, objType);
-						ASSERT(psMessage->psObj, "Viewdata object id %d not found for message %d", objId, id);
-					}
-					else
-					{
-						debug(LOG_ERROR, "Proximity object could not be created (type=%d, player=%d, message=%d)",
-							type, player, id);
-					}
-				}
-				else
-				{
-					VIEWDATA* psViewData = nullptr;
-
-					// Proximity position so get viewdata pointer from the name
-					MESSAGE* psMessage = addMessage(type, false, player);
-
-					if (psMessage)
-					{
-						if (dataType == MSG_DATA_BEACON)
-						{
-							//See addBeaconMessage(). psMessage->dataType is wrong here because
-							//addMessage() calls createMessage() which defaults dataType to MSG_DATA_DEFAULT.
-							//Later when findBeaconMsg() attempts to find a placed beacon it can't because
-							//the dataType is wrong.
-							psMessage->dataType = MSG_DATA_BEACON;
-							Vector2i pos = ini.vector2i("position");
-							int sender = ini.value("sender").toInt();
-							psViewData = CreateBeaconViewData(sender, pos.x, pos.y);
-							ASSERT(psViewData, "Could not create view data for message %d", id);
-							if (psViewData == nullptr)
-							{
-								// Skip this message
-								removeMessage(psMessage, player);
-								continue;
-							}
-						}
-						else if (ini.contains("name"))
-						{
-							psViewData = getViewData(ini.value("name").toWzString());
-							ASSERT(psViewData, "Failed to find view data for proximity position - skipping message %d", id);
-							if (psViewData == nullptr)
-							{
-								// Skip this message
-								removeMessage(psMessage, player);
-								continue;
-							}
-						}
-						else
-						{
-							debug(LOG_ERROR, "Proximity position with empty name skipped (message %d)", id);
-							removeMessage(psMessage, player);
-							continue;
-						}
-
-						psMessage->pViewData = psViewData;
-						// Check the z value is at least the height of the terrain
-						const int height = map_Height(((VIEW_PROXIMITY*)psViewData->pData)->x, ((VIEW_PROXIMITY*)psViewData->pData)->y);
-						if (((VIEW_PROXIMITY*)psViewData->pData)->z < height)
-						{
-							((VIEW_PROXIMITY*)psViewData->pData)->z = height;
-						}
-					}
-					else
-					{
-						debug(LOG_ERROR, "Proximity position could not be created (type=%d, player=%d, message=%d)", type, player, id);
-					}
-				}
-			}
-		}
-		else
-		{
-			// Only load Campaign/Mission messages if a mid-mission save game; always load research messages
-			if (type == MSG_RESEARCH || getSaveGameType() == GTYPE_SAVE_MIDMISSION)
-			{
-				MESSAGE* psMessage = addMessage(type, false, player);
-				ASSERT(psMessage, "Could not create message %d", id);
-				if (psMessage)
-				{
-					psMessage->pViewData = getViewData(ini.value("name").toWzString());
-					ASSERT(psMessage->pViewData, "Failed to find view data for message %d", id);
-				}
-			}
-		}
-		ini.endGroup();
-	}
-	jsDebugMessageUpdate();
-	return true;
-}
-
-// -----------------------------------------------------------------------------------------
-// Load a file from a save game into the psx.
-// This is divided up into 2 parts ...
-//
-// if it is a level loaded up from CD then UserSaveGame will by false
-// UserSaveGame ... Extra stuff to load after scripts
-static bool loadMissionExtras(const char* pGameToLoad, LEVEL_TYPE levelType)
-{
-	char			aFileName[256];
-	UDWORD			fileExten;
-
-	sstrcpy(aFileName, pGameToLoad);
-	fileExten = strlen(pGameToLoad) - 3;
-	aFileName[fileExten - 1] = '\0';
-	strcat(aFileName, "/");
-
-	if (getSaveGameVersion() >= VERSION_11)
-	{
-		//if user save game then load up the messages AFTER any droids or structures are loaded
-		if (getSaveGameType() == GTYPE_SAVE_START || getSaveGameType() == GTYPE_SAVE_MIDMISSION)
-		{
-			//load in the message list file
-			aFileName[fileExten] = '\0';
-			strcat(aFileName, "messtate.json");
-			if (!loadSaveMessage(aFileName, levelType))
-			{
-				debug(LOG_ERROR, "Failed to load mission extras from %s", aFileName);
-				return false;
-			}
-		}
-	}
-
-	return true;
-}
-
 // load up the data for a level
 bool levLoadData(char const *name, Sha256 const *hash, char *pSaveName, GAME_TYPE saveType)
 {
